+++ apps/desktop/src/main.ts
@@ -20,13 +20,7 @@ import {
 } from "electron";
 import type { MenuItemConstructorOptions, OpenDialogOptions } from "electron";
 import type {
-  AuthBearerBootstrapResult,
-  AuthSessionState,
-  AuthWebSocketTokenResult,
   ClientSettings,
-  DesktopSshPasswordPromptRequest,
-  ExecutionEnvironmentDescriptor,
-  DesktopSshEnvironmentTarget,
   DesktopTheme,
   DesktopAppBranding,
   DesktopServerExposureMode,
@@ -65,7 +59,7 @@ import {
   resolveDesktopCoreAdvertisedEndpoints,
   resolveDesktopServerExposure,
 } from "./serverExposure.ts";
-import { DesktopSshEnvironmentManager, resolveRemoteT3CliPackageSpec } from "./sshEnvironment.ts";
+import { DesktopSshEnvironmentBridge, resolveRemoteT3CliPackageSpec } from "./sshEnvironment.ts";
 import { syncShellEnvironment } from "./syncShellEnvironment.ts";
 import { waitForBackendStartupReady } from "./backendStartupReadiness.ts";
 import { getAutoUpdateDisabledReason, shouldBroadcastDownloadProgress } from "./updateState.ts";
@@ -111,14 +105,6 @@ const SET_SAVED_ENVIRONMENT_REGISTRY_CHANNEL = "desktop:set-saved-environment-re
 const GET_SAVED_ENVIRONMENT_SECRET_CHANNEL = "desktop:get-saved-environment-secret";
 const SET_SAVED_ENVIRONMENT_SECRET_CHANNEL = "desktop:set-saved-environment-secret";
 const REMOVE_SAVED_ENVIRONMENT_SECRET_CHANNEL = "desktop:remove-saved-environment-secret";
-const DISCOVER_SSH_HOSTS_CHANNEL = "desktop:discover-ssh-hosts";
-const ENSURE_SSH_ENVIRONMENT_CHANNEL = "desktop:ensure-ssh-environment";
-const FETCH_SSH_ENVIRONMENT_DESCRIPTOR_CHANNEL = "desktop:fetch-ssh-environment-descriptor";
-const BOOTSTRAP_SSH_BEARER_SESSION_CHANNEL = "desktop:bootstrap-ssh-bearer-session";
-const FETCH_SSH_SESSION_STATE_CHANNEL = "desktop:fetch-ssh-session-state";
-const ISSUE_SSH_WEBSOCKET_TOKEN_CHANNEL = "desktop:issue-ssh-websocket-token";
-const SSH_PASSWORD_PROMPT_CHANNEL = "desktop:ssh-password-prompt";
-const RESOLVE_SSH_PASSWORD_PROMPT_CHANNEL = "desktop:resolve-ssh-password-prompt";
 const GET_SERVER_EXPOSURE_STATE_CHANNEL = "desktop:get-server-exposure-state";
 const SET_SERVER_EXPOSURE_MODE_CHANNEL = "desktop:set-server-exposure-mode";
 const GET_ADVERTISED_ENDPOINTS_CHANNEL = "desktop:get-advertised-endpoints";
@@ -177,7 +163,6 @@ function resolvePickFolderDefaultPath(rawOptions: unknown): string | undefined {
 }
 const DESKTOP_LOOPBACK_HOST = "127.0.0.1";
 const DESKTOP_REQUIRED_PORT_PROBE_HOSTS = ["0.0.0.0", "::"] as const;
-const SSH_PASSWORD_PROMPT_TIMEOUT_MS = 3 * 60 * 1000;
 const TITLEBAR_HEIGHT = 40;
 const TITLEBAR_COLOR = "#01000000"; // #00000000 does not work correctly on Linux
 const TITLEBAR_LIGHT_SYMBOL_COLOR = "#1f2937";
@@ -222,12 +207,6 @@ type LinuxDesktopNamedApp = Electron.App & {
   setDesktopName?: (desktopName: string) => void;
 };
 
-interface PendingSshPasswordPrompt {
-  readonly resolve: (password: string | null) => void;
-  readonly reject: (error: Error) => void;
-  readonly timeout: ReturnType<typeof setTimeout>;
-}
-
 let mainWindow: BrowserWindow | null = null;
 let backendProcess: ChildProcess.ChildProcess | null = null;
 let backendPort = 0;
@@ -251,7 +230,6 @@ let restoreStdIoCapture: (() => void) | null = null;
 let backendObservabilitySettings = readPersistedBackendObservabilitySettings();
 let desktopSettings = readDesktopSettings(DESKTOP_SETTINGS_PATH, app.getVersion());
 let desktopServerExposureMode: DesktopServerExposureMode = desktopSettings.serverExposureMode;
-const pendingSshPasswordPrompts = new Map<string, PendingSshPasswordPrompt>();
 
 let destructiveMenuIconCache: Electron.NativeImage | null | undefined;
 const expectedBackendExitChildren = new WeakSet<ChildProcess.ChildProcess>();
@@ -428,7 +406,7 @@ function relaunchDesktopApp(reason: string): void {
           `desktop relaunch backend shutdown warning message=${formatErrorMessage(error)}`,
         );
       })
-      .then(() => desktopSshEnvironmentManager.dispose().catch(() => undefined))
+      .then(() => desktopSshEnvironmentBridge.dispose().catch(() => undefined))
       .finally(() => {
         restoreStdIoCapture?.();
         if (isDevelopment) {
@@ -489,133 +467,6 @@ function getSafeTheme(rawTheme: unknown): DesktopTheme | null {
   }
 
   return null;
-}
-
-function getSafeDesktopSshTarget(rawTarget: unknown): DesktopSshEnvironmentTarget | null {
-  if (typeof rawTarget !== "object" || rawTarget === null) {
-    return null;
-  }
-
-  const target = rawTarget as Partial<DesktopSshEnvironmentTarget>;
-  if (typeof target.alias !== "string" || typeof target.hostname !== "string") {
-    return null;
-  }
-  if (
-    target.username !== null &&
-    target.username !== undefined &&
-    typeof target.username !== "string"
-  ) {
-    return null;
-  }
-  if (target.port !== null && target.port !== undefined && !Number.isInteger(target.port)) {
-    return null;
-  }
-
-  const alias = target.alias.trim();
-  const hostname = target.hostname.trim();
-  if (alias.length === 0 || hostname.length === 0) {
-    return null;
-  }
-
-  return {
-    alias,
-    hostname,
-    username: target.username?.trim() || null,
-    port: target.port ?? null,
-  };
-}
-
-function isLoopbackHostname(hostname: string): boolean {
-  const normalized = hostname
-    .trim()
-    .toLowerCase()
-    .replace(/^\[(.*)\]$/, "$1");
-  return normalized === "127.0.0.1" || normalized === "::1" || normalized === "localhost";
-}
-
-function resolveLoopbackSshHttpUrl(rawHttpBaseUrl: unknown, pathname: string): URL {
-  if (typeof rawHttpBaseUrl !== "string" || rawHttpBaseUrl.trim().length === 0) {
-    throw new Error("Invalid SSH forwarded http base URL.");
-  }
-
-  let baseUrl: URL;
-  try {
-    baseUrl = new URL(rawHttpBaseUrl);
-  } catch {
-    throw new Error("Invalid SSH forwarded http base URL.");
-  }
-
-  if (!isLoopbackHostname(baseUrl.hostname)) {
-    throw new Error("SSH desktop bridge only supports loopback forwarded URLs.");
-  }
-
-  const url = new URL(baseUrl.toString());
-  url.pathname = pathname;
-  url.search = "";
-  url.hash = "";
-  return url;
-}
-
-async function readRemoteFetchErrorMessage(
-  response: Response,
-  fallbackMessage: string,
-): Promise<string> {
-  const text = await response.text();
-  if (!text) {
-    return fallbackMessage;
-  }
-
-  try {
-    const parsed = JSON.parse(text) as { readonly error?: string };
-    if (typeof parsed.error === "string" && parsed.error.trim().length > 0) {
-      return parsed.error;
-    }
-  } catch {
-    // Fall back to the raw text below.
-  }
-
-  return text;
-}
-
-async function fetchLoopbackSshJson<T>(input: {
-  readonly httpBaseUrl: unknown;
-  readonly pathname: string;
-  readonly method?: "GET" | "POST";
-  readonly bearerToken?: unknown;
-  readonly body?: unknown;
-}): Promise<T> {
-  const requestUrl = resolveLoopbackSshHttpUrl(input.httpBaseUrl, input.pathname).toString();
-  const bearerToken =
-    typeof input.bearerToken === "string" && input.bearerToken.trim().length > 0
-      ? input.bearerToken
-      : null;
-
-  let response: Response;
-  try {
-    response = await fetch(requestUrl, {
-      method: input.method ?? "GET",
-      headers: {
-        ...(input.body !== undefined ? { "content-type": "application/json" } : {}),
-        ...(bearerToken ? { authorization: `Bearer ${bearerToken}` } : {}),
-      },
-      ...(input.body !== undefined ? { body: JSON.stringify(input.body) } : {}),
-    });
-  } catch (error) {
-    throw new Error(
-      `Failed to reach SSH forwarded endpoint ${requestUrl} (${error instanceof Error ? error.message : String(error)}).`,
-      { cause: error },
-    );
-  }
-
-  if (!response.ok) {
-    const message = await readRemoteFetchErrorMessage(
-      response,
-      `SSH forwarded request failed (${response.status}).`,
-    );
-    throw new Error(`[ssh_http:${response.status}] ${message}`);
-  }
-
-  return (await response.json()) as T;
 }
 
 async function waitForBackendHttpReady(
@@ -812,54 +663,8 @@ let updateInstallInFlight = false;
 let updaterConfigured = false;
 let updateState: DesktopUpdateState = initialUpdateState();
 
-function rejectPendingSshPasswordPrompts(message: string): void {
-  for (const [requestId, pending] of pendingSshPasswordPrompts) {
-    clearTimeout(pending.timeout);
-    pendingSshPasswordPrompts.delete(requestId);
-    pending.reject(new Error(message));
-  }
-}
-
-async function requestSshPasswordFromRenderer(input: {
-  readonly destination: string;
-  readonly username: string | null;
-  readonly prompt: string;
-}): Promise<string | null> {
-  const window = mainWindow;
-  if (!window || window.isDestroyed()) {
-    throw new Error("T3 Code window is not available for SSH authentication.");
-  }
-
-  const request: DesktopSshPasswordPromptRequest = {
-    requestId: Crypto.randomUUID(),
-    destination: input.destination,
-    username: input.username,
-    prompt: input.prompt,
-  };
-
-  return await new Promise<string | null>((resolve, reject) => {
-    const timeout = setTimeout(() => {
-      pendingSshPasswordPrompts.delete(request.requestId);
-      reject(new Error(`SSH authentication timed out for ${input.destination}.`));
-    }, SSH_PASSWORD_PROMPT_TIMEOUT_MS);
-    timeout.unref();
-
-    pendingSshPasswordPrompts.set(request.requestId, {
-      resolve,
-      reject,
-      timeout,
-    });
-
-    window.webContents.send(SSH_PASSWORD_PROMPT_CHANNEL, request);
-    if (window.isMinimized()) {
-      window.restore();
-    }
-    window.focus();
-  });
-}
-
-const desktopSshEnvironmentManager = new DesktopSshEnvironmentManager({
-  passwordProvider: requestSshPasswordFromRenderer,
+const desktopSshEnvironmentBridge = new DesktopSshEnvironmentBridge({
+  getMainWindow: () => mainWindow,
   resolveCliPackageSpec: () =>
     resolveRemoteT3CliPackageSpec({
       appVersion: app.getVersion(),
@@ -1886,100 +1691,7 @@ function registerIpcHandlers(): void {
     },
   );
 
-  ipcMain.removeHandler(DISCOVER_SSH_HOSTS_CHANNEL);
-  ipcMain.handle(DISCOVER_SSH_HOSTS_CHANNEL, async () =>
-    desktopSshEnvironmentManager.discoverHosts(),
-  );
-
-  ipcMain.removeHandler(ENSURE_SSH_ENVIRONMENT_CHANNEL);
-  ipcMain.handle(
-    ENSURE_SSH_ENVIRONMENT_CHANNEL,
-    async (_event, rawTarget: unknown, rawOptions: unknown) => {
-      const target = getSafeDesktopSshTarget(rawTarget);
-      if (!target) {
-        throw new Error("Invalid desktop SSH target.");
-      }
-
-      const issuePairingToken =
-        typeof rawOptions === "object" &&
-        rawOptions !== null &&
-        "issuePairingToken" in rawOptions &&
-        (rawOptions as { issuePairingToken?: unknown }).issuePairingToken === true;
-
-      return await desktopSshEnvironmentManager.ensureEnvironment(target, {
-        issuePairingToken,
-      });
-    },
-  );
-
-  ipcMain.removeHandler(FETCH_SSH_ENVIRONMENT_DESCRIPTOR_CHANNEL);
-  ipcMain.handle(
-    FETCH_SSH_ENVIRONMENT_DESCRIPTOR_CHANNEL,
-    async (_event, rawHttpBaseUrl: unknown) =>
-      fetchLoopbackSshJson<ExecutionEnvironmentDescriptor>({
-        httpBaseUrl: rawHttpBaseUrl,
-        pathname: "/.well-known/t3/environment",
-      }),
-  );
-
-  ipcMain.removeHandler(BOOTSTRAP_SSH_BEARER_SESSION_CHANNEL);
-  ipcMain.handle(
-    BOOTSTRAP_SSH_BEARER_SESSION_CHANNEL,
-    async (_event, rawHttpBaseUrl: unknown, rawCredential: unknown) =>
-      fetchLoopbackSshJson<AuthBearerBootstrapResult>({
-        httpBaseUrl: rawHttpBaseUrl,
-        pathname: "/api/auth/bootstrap/bearer",
-        method: "POST",
-        body: {
-          credential: rawCredential,
-        },
-      }),
-  );
-
-  ipcMain.removeHandler(FETCH_SSH_SESSION_STATE_CHANNEL);
-  ipcMain.handle(
-    FETCH_SSH_SESSION_STATE_CHANNEL,
-    async (_event, rawHttpBaseUrl: unknown, rawBearerToken: unknown) =>
-      fetchLoopbackSshJson<AuthSessionState>({
-        httpBaseUrl: rawHttpBaseUrl,
-        pathname: "/api/auth/session",
-        bearerToken: rawBearerToken,
-      }),
-  );
-
-  ipcMain.removeHandler(ISSUE_SSH_WEBSOCKET_TOKEN_CHANNEL);
-  ipcMain.handle(
-    ISSUE_SSH_WEBSOCKET_TOKEN_CHANNEL,
-    async (_event, rawHttpBaseUrl: unknown, rawBearerToken: unknown) =>
-      fetchLoopbackSshJson<AuthWebSocketTokenResult>({
-        httpBaseUrl: rawHttpBaseUrl,
-        pathname: "/api/auth/ws-token",
-        method: "POST",
-        bearerToken: rawBearerToken,
-      }),
-  );
-
-  ipcMain.removeHandler(RESOLVE_SSH_PASSWORD_PROMPT_CHANNEL);
-  ipcMain.handle(
-    RESOLVE_SSH_PASSWORD_PROMPT_CHANNEL,
-    async (_event, rawRequestId: unknown, rawPassword: unknown) => {
-      if (typeof rawRequestId !== "string" || rawRequestId.trim().length === 0) {
-        throw new Error("Invalid SSH password prompt id.");
-      }
-      if (rawPassword !== null && typeof rawPassword !== "string") {
-        throw new Error("Invalid SSH password prompt response.");
-      }
-
-      const pending = pendingSshPasswordPrompts.get(rawRequestId);
-      if (!pending) {
-        throw new Error("SSH password prompt is no longer pending.");
-      }
-
-      clearTimeout(pending.timeout);
-      pendingSshPasswordPrompts.delete(rawRequestId);
-      pending.resolve(rawPassword);
-    },
-  );
+  desktopSshEnvironmentBridge.registerIpcHandlers(ipcMain);
 
   ipcMain.removeHandler(GET_SERVER_EXPOSURE_STATE_CHANNEL);
   ipcMain.handle(GET_SERVER_EXPOSURE_STATE_CHANNEL, async () => getDesktopServerExposureState());
@@ -2356,7 +2068,7 @@ function createWindow(): BrowserWindow {
   }
 
   window.on("closed", () => {
-    rejectPendingSshPasswordPrompts(
+    desktopSshEnvironmentBridge.cancelPendingPasswordPrompts(
       "SSH authentication was cancelled because the app window closed.",
     );
     if (mainWindow === window) {
@@ -2450,7 +2162,7 @@ app.on("before-quit", () => {
   clearUpdatePollTimer();
   cancelBackendReadinessWait();
   stopBackend();
-  void desktopSshEnvironmentManager.dispose().catch(() => undefined);
+  void desktopSshEnvironmentBridge.dispose().catch(() => undefined);
   restoreStdIoCapture?.();
 });
 
@@ -2500,7 +2212,7 @@ if (process.platform !== "win32") {
     clearUpdatePollTimer();
     cancelBackendReadinessWait();
     stopBackend();
-    void desktopSshEnvironmentManager.dispose().catch(() => undefined);
+    void desktopSshEnvironmentBridge.dispose().catch(() => undefined);
     restoreStdIoCapture?.();
     app.quit();
   });
@@ -2511,7 +2223,7 @@ if (process.platform !== "win32") {
     writeDesktopLogHeader("SIGTERM received");
     clearUpdatePollTimer();
     stopBackend();
-    void desktopSshEnvironmentManager.dispose().catch(() => undefined);
+    void desktopSshEnvironmentBridge.dispose().catch(() => undefined);
     restoreStdIoCapture?.();
     app.quit();
   });
